# Perf: Optimize `Sqrt` in `Fp2` for all fields