--- a/ecc/bn254/fp/element.go
+++ b/ecc/bn254/fp/element.go
@@ -1302,7 +1302,7 @@ func (z *Element) Sqrt(x *Element) *Element {
 	// q ≡ 3 (mod 4)
 	// using  z ≡ ± x^((p+1)/4) (mod q)
 	var y, square Element
-	y.expBySqrtExp(*x)
+	y.ExpBySqrtExp(*x)
 	// as we didn't compute the legendre symbol, ensure we found y such that y * y = x
 	square.Square(&y)
 	if square.Equal(x) {