--- a/ecc/bw6-633/fp/element.go
+++ b/ecc/bw6-633/fp/element.go
@@ -1901,7 +1901,7 @@ func (z *Element) Sqrt(x *Element) *Element {
 	var one, alpha, beta, tx, square Element
 	one.SetOne()
 	tx.Double(x)
-	alpha.expBySqrtExp(tx)
+	alpha.ExpBySqrtExp(tx)
 
 	beta.Square(&alpha).
 		Mul(&beta, &tx).