# Description

Optimise the square root computations in the quadratic extension `Fp2`. Applications: Hash-to-G2 and G2-points decompression.

## Type of change

- [x] New feature (non-breaking change which adds functionality)


# How has this been tested?

Current Sqrt tests pass.

# How has this been benchmarked?

- BN254:
```
benchmark             old ns/op     new ns/op     delta
BenchmarkE2Sqrt-8     42712         11773         -72.44%
```

- BLS12-381:
```
benchmark             old ns/op     new ns/op     delta
BenchmarkE2Sqrt-8     118420        28176         -76.21%
```

- BLS12-377:
TODO

# Checklist:

- [ ] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I did not modify files generated from templates
- [ ] `golangci-lint` does not output errors locally
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules